requirements.txt
tweepy==4.9.0
requests==2.26.0
Pillow==9.1.0